+++ src/Http/Resources/ResourceCollection.php
@@ -2,8 +2,11 @@
 
 namespace DoubleThreeDigital\Runway\Http\Resources;
 
+use DoubleThreeDigital\Runway\Fieldtypes\BelongsToFieldtype;
+use DoubleThreeDigital\Runway\Fieldtypes\HasManyFieldtype;
 use DoubleThreeDigital\Runway\Runway;
 use Illuminate\Http\Resources\Json\ResourceCollection as LaravelResourceCollection;
+use Illuminate\Support\Str;
 use Statamic\Facades\Action;
 use Statamic\Facades\User;
 
@@ -62,6 +65,27 @@ class ResourceCollection extends LaravelResourceCollection
                     }
 
                     if ($this->runwayResource->blueprint()->hasField($key)) {
+                        // If we've eager loaded in relationships, just pass in the model
+                        // instance. We can save extra queries this way.
+                        if ($this->runwayResource->blueprint()->field($key)->fieldtype() instanceof BelongsToFieldtype) {
+                            // TODO: refactor this to use the 'magic relation guessing' from the
+                            // Resource's `eagerLoadingRelations` method.
+                            $relationName = str_replace('_id', '', $key);
+                            $relationName = Str::camel($relationName);
+
+                            if ($record->relationLoaded($relationName)) {
+                                $value = $record->$relationName;
+                            }
+                        }
+
+                        if ($this->runwayResource->blueprint()->field($key)->fieldtype() instanceof HasManyFieldtype) {
+                            $relationName = $key;
+
+                            if ($record->relationLoaded($relationName)) {
+                                $value = $record->$relationName;
+                            }
+                        }
+
                         $row[$key] = $this->runwayResource->blueprint()->field($key)->setValue($value)->preProcessIndex()->value();
                     }
                 }